Fixed null pointer exception and added user input validation.
There was an error in the last commit where objects were being called
before their initialization. Their order has been fixed and the
exception does not trigger. (In fact, the last commit broke the ability
to add new deck binders on the fly.)

Dialog boxes are in place to help the user through adding a new deck to
a particular deck binder along with validating their input. There is
still no input verification in place.