Deck Moving bug fix + Preview DBP fix
Fixed using the shortcut Shft + UP/DOWN to move decks up and down in the
order they appear in when the comboBox's drop down menu is visible.
Seems to have been broken since inception. Couldn't find a time that the
order changes were saved and usable in the next opening of the
application.

Fixed the DBP in the Preview panel of DBEditDialog showing that odd
colored border effect after selecting an option. + the DBP starts off
with the correct style now.
-There is still a bug where changing the order will /change the order/
but the style of the combobox won't update/will use the old style
anyways.