Fixed "Add new Deck Binder" dialog bugs
When clicking "Cancel" or the close button, the window would not delete
the temporary Deck Binder. This caused an unnamed Deck Binder to show up
when restarting the program.

The base DeckBinderDialog now runs through a new canceAction whether the
user clicks "Cancel" or the close button, ensuring the deletion of the
temporary Deck Binder.