Fixed bugs with deck styles not applying themselves to the CBs
Fixed Bug - Selecting "add new deck" and then hitting cancel makes first
deck in the comboBox appear but no style is applied to it.

Fixed Bug - Selecting "add new deck" and then clicking "done" makes that
deck appear but without any style effects appearing.

Performance - Fixed style being applied twice per item seleciton.
Program only applies the style of the item that is being selected now,
instead of applying the style of the item that was deselected and then
applying the style of the item that was selected.